+++ api/oc_discovery.c
@@ -1027,7 +1027,7 @@ oc_create_discovery_resource(int resource_idx, size_t device)
 
 #ifdef OC_CLIENT
 oc_discovery_flags_t
-oc_discovery_process_payload(const uint8_t *payload, int len,
+oc_discovery_process_payload(const uint8_t *payload, size_t len,
                              oc_client_handler_t client_handler,
                              const oc_endpoint_t *endpoint, void *user_data)
 {
